Add missing links for lessons 6-10 to TOC (#84)
* Add missing lessons to TOC

Tutorial #3 has now 10 lessons. Provide links to them. This has to be done for all lessons in the tutorial on regression modelling (copy & paste).

I will provide the next few days also the changes for the book to this issue.

NOTE: I could not manage the R code for the emoji of lesson 10 and have it included as inline icon. `r emo:ji("ten")` displays a 10 o'clock watch in my installation.

NOTE: Lesson 6 cannot be called because of a missing graph in a knitr::include_graphics() command. (lines 457-458). I think I have found the problem: graphic of chunk "mpg-data-parallel" has to be saved + called. If this is correct a PR to this issue will follow.

* Update 03-model/01-lesson/03-01-lesson.Rmd

Co-authored-by: Mine Cetinkaya-Rundel <cetinkaya.mine@gmail.com>