Correct variable names and adapt exercise (#109)
It is not clear which variable to use for the last exercise (chunk ex8). It says, "Create a visualization of the relationship between at bats and obp…" but then all the following commands and text passages refer to slg and obp.

I assume that the relationship between `at_bat` and `obp` is meant. I am not sure if this makes sense because I have no idea about baseball.

[From a cross-cultural perspective, I think that baseball is not a good example. In some parts of the world, this sport is not popular, and people (like me) even know the rules of the game and therefore cannot understand the different variables of the dataset `mlbbat10`.]

closes #92