03-06 Lesson: Several adaptions and fixing some typos (#111)
* show r chunk for data_space calculation

Otherwise "data_space" in the code chunk of the section "Parallel lines on the scatterplot" would not be understandable.

* change name of numeric explanatory variable

Name is `displ` not `displacement`.

* Make it clear that the text is now referring to the `mariokart` data.

I recommend this relatively small adaption so that the example is better understandable. It would be beneficial to revise the lesson: There are several changes between the explanation model (mpg) and the exercise model (`mariokart`). Besides that, some text passages are redundant; the lesson lacks signals whenever the change between the two models occurs. It is difficult for students to follow the text as the corresponding data is not easily accessible but is hidden in another section of this lesson. Also (as a general recommendation), I would suggest - in addition to the glimpse() command - always providing links to the data set explanation page (in this case, a link to https://www.openintro.org/data/index.php?data=mariokart).

My strategy with these PRs is to make the most necessary and easy changes and overhaul the whole lesson later.

* Delete one of the two task descriptions.

There are two task descriptions with essentially the same advice. I recommend deleting the first one as the second seems a little bit clearer worded.

* change 'set of slides' to 'sections'.

* Add references and links to data set descriptions

Add also missing names of variables and correct spelling of explanatory variables x1 & x2.