lesson 03 07 adaption (#112)
* New numbering after structural change

* 2 Typos in section "Model Fit"

- add `$` before and behind to 'number of explanatory variables p'

- delete spaces after and before adjusted R^2 formula (last formula in section

* delete several times `.` after $R^2$ when the sentence is not finished

* typo: '…as new variable[s] are added…'

* Adapt the model fit exercise

- generate a new chapter 'Your turn'
- remember that we are continuing with the `mariokart` example
- condense the two task description to one  task description
- adapt the hints accordingly (adding `n` in the mutate call)

* correct variable name in formula

add twice `\` to the variable name `is_newer` in the formula to get the underline instead the subscript

* Explain how the second slope is calculated

For students it might not clear how to get  -3.46 mpg per liter for newer cars.

* Put hint after question.

I have used the HTML element `details` because it generates an accordion starting with a triangle followed by the word "Details".

BTW: This is also the solution for using markdown hints together with rmarkdown hints! You need to write into the markdown document the text between the HTML tag <details>. see, e.g., https://www.w3schools.com/TAGS/tag_details.asp

You can even change the word "Details" for instance to "Hint". In this case you would need to write as normal text (markdown) into the document:

<displays>
<summary><strong>Hint</strong><summary>
</display>

Because of a special CSS in `learnr` documents displaying the triangle in `learnr` documents (not in most other documents), you would need to add a new CSS rule:

details > summary {
  display: list-item;
}

You could also provide changing text with open and closed accordions with the following CSS rule for example:

details > summary::before {
    content: '▶︎ Click here to show the solution:';
}

details[open] > summary::before {
  content: '︎▼ Click here to hide the solution:';
}

Then you would add in the R Markdown document:

<details>
   <summary>
      <strong>Show solution</strong>
   </summary>
</details>

I would suggest using the standard version with "Details" as it is the most flexible way, differentiating with the markdown learner hints (which nevertheless could be used) and is still self-explanatory.

* resolve conflict between div hint and task solution

You cannot use both: markdown and rmarkdown hint at the same time. Instead of providing the hint text as an RMarkdown comment, I suggest using <details> as explained in the previous commit.

* To see the solution change div hint to details

add second bullet point for the second part of task description

* change structure and insert calculations

The two sections, "SAT scores and teacher salary" and "Consequences of Simpson's paradox," were missing calculations and were not understandable.

I had to restructure the content and add the corresponding code chunks. The original setup chunk had to be changed too. Otherwise, there would not have been a difference with and without binned salary.

PLEASE CHECK MY CODE CHUNKS! I get different values: Instead of -5.02, my value is 5.54, and instead of 1.84, I got 3.09. MAYBE MY CALCULATION IS NOT CORRECT?

* section restructured

- syncronize text with visualization
- change div hint to details accordion

* Adapt links to new structure

Maybe you could find time to upload these changes to shinyapps.io? So that I could see if all link changes are correct. It gets complicated to remember which links and TOCs are already valid and which are not.

* change all 'Details" to '**Hint**'

And add CSS file with the necessary rule to revert `learnr`-template specific to standard.

(Even if the CSS rule would be equal for all lessons, I think it is better to provide the CSS file for every lesson and make every lesson independent from the others.)