Chained appropriate "exercise.setup"s for many exercises and fixed various typos and problems with code/exercises not displaying.  (#154)
* Restructure section with additional information

- I think it would be helpful to take every lesson as an independent lection. Therefore all starting information has to be applied again.

- I changed the chunk option 'eval=FALSE' to 'exercise=TRUE' because students

(a) should see the results of the code explanation and

(b) should be engaged with some interaction

To prevent conflicts with the following exercises' names, I also had to rename the chunk labels. Where available, I construct them with "ex0[serial number]_[old-label]."

* add explanation of colored 2D data space

- merge two headers, "Not longer a 2D problem" and "Data space is 3D" to "Data space is 3D."

- convert R chunk to exercise setup chunk (`data_space` is needed later)

- explain what the plot is doing

* add more detailed explanations

Because of the sparse explication, I believe that the code chunks are not understandable for students. At least it was challenging for me. So I suggest providing more explanation and restructuring the text. (As I am not a domain expert and not an English native, please correct where it is necessary!)

I also had to revise the exercise chunk labels before and after this changed section. It turned out that shared exercise setup chunks are not appropriate. When developing step-by-step `learnr` exercises, chained setup chunks are necessary. See for details: https://rstudio.github.io/learnr/exercises.html#Exercise_Setup_Chunks

I (will) establish two sets of exercise setup labels: one for the `babies` and the other for the coming `mariokart` data set exercises. Each one follows the rule: <ex[number]-[dataset]-[name]>. The resulting labels may seem too long, but I think they may come in handy whenever one rearranges the code.

* add detailed explanations

I tried to explain the (complex) procedure with many new commands similar to the previous tiled visualization.

I deleted the sentence "The x, y, and plane objects were created in code that we aren't showing here due to its complexity." because the code is shown. I tried to explain the code snippets as well.

Again: As I am not a domain expert and not an English native, please correct where necessary!)

* add intro for the upcoming lesson

If we want to provide every lesson as an independent tutorial, some extra information would be helpful. (As I already mentioned twice, please correct (delete, add) my text suggestions where necessary!)

* Some addition and re-arrangement

- Add new section title "Your turn"
- Describe the general task
- Add glimpse chunks
- Add advice to look at data set at OpenIntro page.

* display preparation code and fix errors

- fix typo: `marioKart` instead `mariokart`
- load package `plotly`
- remove markdown hint (destroys solution but is also unnecessary/ redundant
- display preparation code: Hidden code is not helpful for learning purposes. In general: The example is very complex and has many details which are only explained en-passant.

* turn presented R code to exercise code

+ In addition to "exercise=TRUE" I had to add exercise.setup to build up chained setup chunks.

+ Rename the exercise chunks accordingly to the babies example labels

+ Turn the last R code chunk to an exercise with gaps.

* Reorder some pragraphs

Restructure some paragraphs after changing from presentation to exercise code.

* Change presented code to exercise

* integrate hint into question

convert external hint into a message of the quiz, which will be displayed when the answer is chosen

* turn standard r chunks to exercise chunks

also
- restructure text
- load packages for the plotly chunk

* integrate hint into MC and delete duplicated "Correct!" message

* delete exercise=TRUE from solution chunk

* Change label of plotly exercise and tix some typos

* delete two R code chunks

These two code chunks are redundant (duplicates). It seems they are leftovers form the reorganization of the lesson/chapter,

* Convert R Code to exercise code and restructure the explaing text

- add library(plotly) otherwise the code would not run
- integrate markdown hint into the solution

* Convert R Code chunks to exercise chunks and restructure text accordingly

* delete redundant R code

* Convert R code into exercise and integrate hint into anser of MC

* set tutorial option 'exercise.blanks'

For this new option you have to download the new `learnr` GitHub version

* add knitr::opts_chunk `verbose=FALSE`

Sometimes I got with plotly the error message "verbose must be TRUE or FALSE". I tried to prevent it with this options. But it does not help (always) ???

* Convert text to MC

- integrate hint into MC
- renumber all following labels for MCs

Please check if answer option is the correct one, because there was no solution in the text passage.

* Chained appropriate "exercise.setup"s for the exercises using the "noise" dataset in order to avoid the "object 'noise' not found" errors that were occuring (issue #153).

* Fixed #152.

Code was not showing up in ex2-setup because it was not itself an exercise.

* 02-01-lesson.Rmd would not run. Error caused by misidentified id_cols in the calls to pivot_wider().

* Fixes #151 typo.

* Fixes #150, #149, #148, and #147.
Mainly, these issues stemmed from problems with chunks not being displayed by default and exercises not inheriting appropriate setups.

* Fixes #144 by modifying the new figure for the "CIs and confidence level" question

Fixes #146 and fixes #145 typo.

* Fixes #138.

Pie chart was not constructed properly.

* Fixes #137.

* Fixes #136

* Added pacman::p_load to setup chunk to install any missing packages needed for the tutorials. This will allow users who download the code to run the documents from within Rstudio without worrying about not having the needed packages.

Issue #6 would address this better, but this will make things a little smoother in the meantime.

Co-authored-by: petzi53 <petzi53@gmail.com>
Co-authored-by: Mine Cetinkaya-Rundel <cetinkaya.mine@gmail.com>